+++ lib/Cake/Test/Case/I18n/L10nTest.php
@@ -36,6 +36,19 @@ class L10nTest extends CakeTestCase {
 	}
 
 /**
+ * tearDown method
+ *
+ * @return void
+ */
+	public function tearDown() : void {
+		parent::tearDown();
+
+		if (isset($_SERVER['HTTP_ACCEPT_LANGUAGE'])) {
+			unset($_SERVER['HTTP_ACCEPT_LANGUAGE']);
+		}
+	}
+
+/**
  * testGet method
  *
  * @return void