+++ lib/Cake/Test/Case/Utility/XmlTest.php
@@ -607,11 +607,17 @@ XML;
  * @return void
  */
 	public function testFromArrayFail($value) {
+		set_error_handler(function($errno, $errstr) {
+			return true;
+		}, E_WARNING | E_USER_WARNING);
+
 		try {
 			Xml::fromArray($value);
 			$this->fail('No exception.');
 		} catch (Exception $e) {
 			$this->assertTrue(true, 'Caught exception.');
+		} finally {
+			restore_error_handler();
 		}
 	}